+++ src/Sonos-Commands.js
@@ -249,7 +249,7 @@ module.exports = {
         })
     }
     if (snapshot.positionInfo.RelTime && snapshot.positionInfo.TrackDuration !== '0:00:00') {
-      node.debug('Setting back time to >>', JSON.stringify(snapshot.positionInfo.RelTime))
+      node.debug('Setting back time to >>' + JSON.stringify(snapshot.positionInfo.RelTime))
       await membersAsPlayerPlus[coordinatorIndex].avTransportService().Seek({ InstanceID: 0, Unit: 'REL_TIME', Target: snapshot.positionInfo.RelTime })
         .catch(reason => {
           node.debug('Reverting back track time failed, happens for some music services (radio or stream).')