+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [3.0.2] 2020-04-23
+
+### Changed
+
+- Control Player: bug fix serialnum
+
+- Helper, error codes: new error handling
+
 ## [3.0.1] 2020-04-18
 
 ### Changed